Move TransformedModule to bundler::load
